requirements.txt
opencv-python
pytesseract
numpy
pandas
Pillow
pdf2image
matplotlib
python-docx
PyYAML
pytest
easyocr
scikit-image
reportlab
tqdm